onboarding: move other screens
